onlineshopping

+++ controllers/user_controller.js
@@ -1,13 +1,23 @@
 const express =require("express")
 const router = express.Router()
 const bcrypt = require('bcrypt')
-const db=require("./../db")
+const db=require("./../database")
 
 // const ensureLoggedIn=require("./../middlewares/ensure_logged_in")
 // const viewHelpers = require("./../middlewares/view_helpers")
 
 
+router.get('/users', (req, res)=>{
 
+    const sql = "select * from users;"
+
+    db.query(sql, (err, dbRes) =>{
+        // console.log(dbRes.rows)
+        const users = dbRes.rows
+        // res.render("home", {dishes: dishes,email: req.session.email})
+        res.render("allusers", {users: users})
+    })
+})
 
 router.get('/users/new',(req,res)=>{
 
@@ -26,7 +36,7 @@ router.post('/users',(req, res)=>{
         bcrypt.hash(plainTextPassword, salt, (err, digestedPassword)=>{
 
         console.log(digestedPassword)
-        const sql = `insert into users (email, password_digest) values($1,'$2);`
+        const sql = `insert into users (email, password_digest) values($1,$2);`
 
         db.query(sql,[email,digestedPassword],(err, dbRes)=>{
 
@@ -40,6 +50,36 @@ router.post('/users',(req, res)=>{
 })
 })
 
+router.get('/users/:id/edit',(req, res)=>{
+    console.log(req.params.id)
+    if (req.params.id!==1){
+        const userId = req.params.id
+        const sql = `select * from users where id = ${userId};`
+
+        db.query(sql, (err, dbRes) =>{
+        // console.log(dbRes)
+            const user = dbRes.rows[0]
+        
+        res.render("edit_user", {user: user})
+        })
+    } else if (req.params.id==1) {
+
+        res.render('admin')
+    }
+})
+
+router.put('/users/:id', (req,res)=>{
+
+    console.log(req.body.email)
+    const sql = `update users set email = $1, address = $2 where id = $3;` 
+
+    db.query(sql, [req.body.email, req.body.address, req.params.id],( err, dbRes)=>{
+        res.redirect('/users')
+        
+
+    })
+
+})
 
 router.get("/users/:id",(req, res)=>{
 
@@ -58,24 +98,28 @@ router.get("/users/:id",(req, res)=>{
         
         }
     })
-    res.render("user_detail", {orders})
+    res.render("userDetail", {orders})
 
 })
 
 
 router.delete("/users/:user_id", (req, res) => {
-    // console.log(req.body.dish_id)
 
-    const sql = `delete from users where id =${req.params.user_id};`
+    if (req.params.user_id!==1){
+        const sql = `delete from users where id =${req.params.user_id};`
 
-    db.query(sql, (err, dbRes)=> {
-        // isLoggedIn()=false
-        req.session.destroy(()=>{
+        db.query(sql, (err, dbRes)=> {
+            // isLoggedIn()=false
+            // req.session.destroy(()=>{
 
-            res.redirect("/login")
+                res.redirect("/users")
+            // })
+
         })
+    }else {
 
-    })
+        res.redirect("/admin")
+    }
 
 })
